if else

See merge request sangnier/codelab3!81

+++ src/view/langage/DeletePanel.java
@@ -24,7 +24,10 @@ public class DeletePanel extends JPanel implements IMouseReactive {
 
     public void onRelease(InstructionPanel source) {
         if (source != null)
-            source.delete();
+            if (source.getInstruction() != null)
+                source.delete();
+            else 
+                source.dehighlight();
     }
 
     public String getDestType() {
